feat(apps,fixtures): apps ships its testing fixtures, and the twice-written model and bridge helpers get one private home (#test-kit)
Two additive surfaces, three suites migrated onto them as proof.

`@vendoai/apps` gains the `./testing` subpath. `src/testing/` was already
compiled into `dist/testing/` and already inside the published `dist`
files-entry — it was shipping to npm and reachable by nobody, and
`scripted-model.test.ts` has been claiming `@vendoai/apps/testing` resolves
since the day it was written. It did not. Same posture as the
`./adapter-conformance` subpath this package already publishes.

`@vendoai-fixtures/test-kit` is a new private fixture package carrying the two
helpers that are genuinely written more than once and could not be written
differently without being wrong: the scripted `LanguageModel` and its stream
parts (13 independent copies of `ZERO_USAGE` alone, across packages/,
fixtures/, corpus/ and examples/), and the node:http-to-Web-Request adapter
(two byte-identical copies in fixtures/mcp-e2e). It has no `.` export on
purpose — `./stream-turns` needs `ai` and `./node-bridge` needs nothing, and a
barrel would make every consumer of either pay for both.

It lives under fixtures/ rather than packages/ deliberately. The dependency
guard requires every @vendoai/* import in a package's src to be a real
dependency, devDeps explicitly not counting, so a scoped kit in packages/ would
have to enter a published package's dependency list to be consumable — a
private test package in a shipped tarball's manifest. Under fixtures/ with the
existing @vendoai-fixtures/ scope it is a devDependency and nothing else, and
scripts/dependency-guard.mjs is untouched (no collision with #983).

Migrated, each run green end to end:
- fixtures/integration/src/harness.ts takes the turn helpers and re-exports
  them, so its 19 journeys are unchanged at the call site (51 passed).
- fixtures/mcp-e2e's two copies of `bridge` become one `nodeBridge` (19 passed).
- fixtures/automations-e2e's two hand-rolled scripted models become
  `scriptedLanguageModel` from `@vendoai/apps/testing` — one of them carried
  the comment "Local copy for the same reason ladder.e2e.test.ts keeps one: the
  apps package's own test double is internal to that package", which is the
  subpath's whole justification, written down before it existed (69 passed).

What was NOT promoted, on purpose: the guard double (`testGuard`/`boundRegistry`,
duplicated between harnesses and vendo). @vendoai/guard ships a real
`createGuard` and packages/guard already drives it against a real store, so a
shared fake standing in for a runnable seam is the mock farm CLAUDE.md forbids —
and both existing copies carry a written decision against sharing them. Same for
`tempStore`: forty copies, every one inside packages/vendo, which is one
package's own helper written forty times rather than a shared one.
