+++ fixtures/integration/package.json
@@ -23,6 +23,7 @@
   "devDependencies": {
     "@modelcontextprotocol/sdk": "^1.29.0",
     "@types/node": "^22.0.0",
+    "@vendoai-fixtures/test-kit": "workspace:*",
     "ai": "^6.0.0",
     "fflate": "^0.8.2",
     "typescript": "^5.6.0",
