+++ fixtures/integration/src/harness.ts
@@ -29,7 +29,16 @@ import type { AppDocument, Principal, ToolRegistry } from "@vendoai/core";
 import { createMcpDoor, type AppsPort, type HostOAuthAdapter, type McpDoor } from "@vendoai/mcp";
 import { createStore, type VendoStore } from "@vendoai/store";
 import { createVendo, type CreateVendoConfig, type Vendo } from "@vendoai/vendo/server";
-import { MockLanguageModelV3, simulateReadableStream } from "ai/test";
+import {
+  scriptedModel,
+  textTurn,
+  toolCallTurn,
+  ZERO_USAGE,
+  type LanguageModelV3Prompt,
+  type LanguageModelV3StreamPart,
+  type ScriptedModel,
+  type ScriptedTurn,
+} from "@vendoai-fixtures/test-kit/stream-turns";
 import type { LanguageModel } from "ai";
 
 export const fixtureBaseUrl = (): string => inject("fixtureBaseUrl");
@@ -59,45 +68,23 @@ export const BOB: Principal = { kind: "user", subject: "user_bob" };
 export const WIRE_BASE = "/api/vendo";
 
 // ---------------------------------------------------------------------------
-// Scripted LanguageModel — the chat-e2e technique. ONE model instance drives
-// BOTH the agent loop (doStream) AND the apps generation engine (doGenerate via
-// generateText); they share one FIFO queue of turns, so a journey scripts turns
-// in the exact order the composed system will consume them.
+// Scripted LanguageModel — the chat-e2e technique, now shared with every other
+// suite that scripts a model (`@vendoai-fixtures/test-kit/stream-turns`).
+// Re-exported here so
+// this harness stays the single import every journey in this package reaches
+// for.
 // ---------------------------------------------------------------------------
 
-type LanguageModelV3Prompt = Parameters<MockLanguageModelV3["doStream"]>[0]["prompt"];
-type LanguageModelV3StreamPart = Awaited<
-  ReturnType<MockLanguageModelV3["doStream"]>
->["stream"] extends ReadableStream<infer Part> ? Part : never;
-type LanguageModelV3GenerateResult = Awaited<ReturnType<MockLanguageModelV3["doGenerate"]>>;
-type LanguageModelV3Content = LanguageModelV3GenerateResult["content"][number];
-
-export const ZERO_USAGE = {
-  inputTokens: { total: 0, noCache: 0, cacheRead: 0, cacheWrite: 0 },
-  outputTokens: { total: 0, text: 0, reasoning: 0 },
-} as const;
-
-/** A plain assistant text turn (agent doStream). */
-export function textTurn(text: string, id = "text_1"): LanguageModelV3StreamPart[] {
-  return [
-    { type: "text-start", id },
-    { type: "text-delta", id, delta: text },
-    { type: "text-end", id },
-    { type: "finish", usage: ZERO_USAGE, finishReason: { unified: "stop", raw: undefined } },
-  ];
-}
-
-/** An agent turn that calls one tool (agent doStream). */
-export function toolCallTurn(
-  toolName: string,
-  input: unknown,
-  toolCallId = "call_1",
-): LanguageModelV3StreamPart[] {
-  return [
-    { type: "tool-call", toolCallId, toolName, input: JSON.stringify(input) },
-    { type: "finish", usage: ZERO_USAGE, finishReason: { unified: "tool-calls", raw: undefined } },
-  ];
-}
+export {
+  scriptedModel,
+  textTurn,
+  toolCallTurn,
+  ZERO_USAGE,
+  type LanguageModelV3Prompt,
+  type LanguageModelV3StreamPart,
+  type ScriptedModel,
+  type ScriptedTurn,
+};
 
 /** A generation-engine turn: the apps engine reads this through doGenerate and
  * parses the emitted text as CREATE/EDIT-dialect JSON. Pass the object; it is
@@ -156,60 +143,11 @@ export function screenAgentCreateTurns(dialect: string): LanguageModelV3StreamPa
   ];
 }
 
-/**
- * One scripted turn: the chunks, or a function of the prompt that was handed to
- * the model.
- *
- * The function form exists for the ids a journey cannot know in advance — the
- * front door mints an app's id at request time and writes it into the screen
- * agent's brief, so a turn that has to NAME that app (`validate({ appId })`)
- * reads it off the brief exactly as the model it stands in for would.
- */
-export type ScriptedTurn =
-  | LanguageModelV3StreamPart[]
-  | ((prompt: LanguageModelV3Prompt) => LanguageModelV3StreamPart[]);
-
 /** The app id the composed server put in the prompt it just sent, if any. */
 export function appIdInPrompt(prompt: LanguageModelV3Prompt): string {
   const found = /app_[0-9a-f-]{8,}/.exec(JSON.stringify(prompt));
   if (found === null) throw new Error("no app id in the prompt the model was handed");
   return found[0];
-}
-
-export type ScriptedModel = MockLanguageModelV3 & { prompts: LanguageModelV3Prompt[] };
-
-export function scriptedModel(turns: readonly ScriptedTurn[]): ScriptedModel {
-  const remaining = turns.map((turn) => (typeof turn === "function" ? turn : [...turn]));
-  const prompts: LanguageModelV3Prompt[] = [];
-  const shift = (prompt: LanguageModelV3Prompt): LanguageModelV3StreamPart[] => {
-    prompts.push(structuredClone(prompt));
-    const turn = remaining.shift();
-    if (turn === undefined) throw new Error("scripted model exhausted");
-    const chunks = typeof turn === "function" ? turn(prompt) : turn;
-    return chunks;
-  };
-  const model = new MockLanguageModelV3({
-    doStream: async (request) => ({ stream: simulateReadableStream({ chunks: shift(request.prompt) }) }),
-    doGenerate: async (request): Promise<LanguageModelV3GenerateResult> => {
-      const chunks = shift(request.prompt);
-      const finish = chunks.find((part) => part.type === "finish");
-      const content: LanguageModelV3Content[] = [];
-      const text = chunks
-        .filter((part): part is Extract<LanguageModelV3StreamPart, { type: "text-delta" }> => part.type === "text-delta")
-        .map((part) => part.delta)
-        .join("");
-      if (text.length > 0) content.push({ type: "text", text });
-      for (const part of chunks) if (part.type === "tool-call") content.push(structuredClone(part));
-      return {
-        content,
-        finishReason: finish?.finishReason ?? { unified: "stop", raw: undefined },
-        usage: finish?.usage ?? ZERO_USAGE,
-        warnings: [],
-      };
-    },
-  }) as ScriptedModel;
-  model.prompts = prompts;
-  return model;
 }
 
 // ---------------------------------------------------------------------------
