+++ fixtures/mcp-e2e/package.json
@@ -24,6 +24,7 @@
     "@modelcontextprotocol/sdk": "^1.29.0",
     "@types/node": "^22.0.0",
     "@types/pg": "^8.11.10",
+    "@vendoai-fixtures/test-kit": "workspace:*",
     "pg": "^8.22.0",
     "typescript": "^5.6.0",
     "vitest": "^2.1.0"
