+++ fixtures/mcp-e2e/src/umbrella-hookup.e2e.test.ts
@@ -13,9 +13,10 @@
  * loopback origin, so the door and the host API are genuinely separate origins.
  */
 import { mkdir, mkdtemp, rm, writeFile } from "node:fs/promises";
-import { createServer, type IncomingMessage, type ServerResponse } from "node:http";
+import { createServer } from "node:http";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
+import { nodeBridge } from "@vendoai-fixtures/test-kit/node-bridge";
 import { type Principal } from "@vendoai/core";
 import { VENDO_TOOLS_FORMAT } from "@vendoai/actions";
 import type { HostOAuthAdapter } from "@vendoai/mcp";
@@ -128,7 +129,7 @@ async function createUmbrella(
   // loopback origin. Pass that door origin explicitly via mcp.baseUrl so
   // discovery advertises the URL the SDK client can actually reach.
   let handler: Vendo["handler"] | undefined;
-  const httpServer = createServer((req, res) => void bridge(req, res, (request) => handler!(request)));
+  const httpServer = createServer((req, res) => void nodeBridge(req, res, (request) => handler!(request)));
   await new Promise<void>((resolve, reject) => {
     httpServer.once("error", reject);
     httpServer.listen(0, "127.0.0.1", () => {
@@ -431,34 +432,3 @@ describe("umbrella hookup — createVendo({ mcp: true }) mounts the door", () =>
     }
   });
 });
-
-async function bridge(
-  req: IncomingMessage,
-  res: ServerResponse,
-  handler: (request: Request) => Promise<Response>,
-): Promise<void> {
-  try {
-    const chunks: Buffer[] = [];
-    for await (const chunk of req) chunks.push(Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk));
-    const host = req.headers.host ?? "127.0.0.1";
-    const headers = new Headers();
-    for (const [name, value] of Object.entries(req.headers)) {
-      if (value === undefined) continue;
-      headers.set(name, Array.isArray(value) ? value.join(", ") : value);
-    }
-    const body = chunks.length === 0 ? undefined : Buffer.concat(chunks);
-    const request = new Request(`http://${host}${req.url ?? "/"}`, {
-      method: req.method,
-      headers,
-      ...(body === undefined ? {} : { body }),
-    });
-    const response = await handler(request);
-    res.statusCode = response.status;
-    response.headers.forEach((value, name) => res.setHeader(name, value));
-    res.end(Buffer.from(await response.arrayBuffer()));
-  } catch (error) {
-    res.statusCode = 500;
-    res.setHeader("content-type", "text/plain");
-    res.end(error instanceof Error ? error.message : "umbrella bridge failed");
-  }
-}
